Implement arguments object.
