+++ core/src/avm1/object.rs
@@ -1,4 +1,5 @@
 use crate::avm1::{ActionContext, Avm1, Value};
+use crate::avm1::function::{Executable, NativeFunction, Avm1Function};
 use crate::display_object::DisplayNode;
 use crate::tag_utils::SwfSlice;
 use core::fmt;
@@ -6,82 +7,6 @@ use gc_arena::{GcCell, MutationContext};
 use std::collections::hash_map::Entry;
 use std::collections::HashMap;
 use std::mem::replace;
-
-pub type NativeFunction<'gc> = fn(
-    &mut Avm1<'gc>,
-    &mut ActionContext<'_, 'gc, '_>,
-    GcCell<'gc, Object<'gc>>,
-    &[Value<'gc>],
-) -> Value<'gc>;
-
-/// Represents a function defined in the AVM1 runtime.
-#[derive(Clone)]
-struct Avm1Function {
-    /// The file format version of the SWF that generated this function.
-    swf_version: u8,
-
-    /// A reference to the underlying SWF data.
-    data: SwfSlice,
-    
-    /// The name of the function, if not anonymous.
-    name: Option<String>,
-
-    /// The names of the function parameters.
-    params: Vec<String>,
-}
-
-impl Avm1Function {
-    pub fn new(swf_version: u8, actions: SwfSlice, name: &str, params: &[&str]) -> Avm1Function {
-        let name = match name {
-            "" => None,
-            name => Some(name.to_string())
-        };
-
-        Avm1Function {
-            swf_version: swf_version,
-            data: actions,
-            name: name,
-            params: params.into_iter().map(|s| s.to_string()).collect()
-        }
-    }
-}
-
-/// Represents a function that can be defined in the Ruffle runtime or by the
-/// AVM1 bytecode itself.
-#[derive(Clone)]
-enum Executable<'gc> {
-    /// A function provided by the Ruffle runtime and implemented in Rust.
-    Native(NativeFunction<'gc>),
-
-    /// ActionScript data defined by a previous action.
-    Action(Avm1Function)
-}
-
-impl<'gc> Executable<'gc> {
-    /// Execute the given code.
-    /// 
-    /// Execution is not guaranteed to have completed when this function
-    /// returns. If on-stack execution is possible, then this function returns
-    /// a return value you must push onto the stack. Otherwise, you must
-    /// create a new stack frame and execute the action data yourself.
-    pub fn exec(&self, avm: &mut Avm1<'gc>, ac: &mut ActionContext<'_, 'gc, '_>, this: GcCell<'gc, Object<'gc>>, args: &[Value<'gc>]) -> Option<Value<'gc>> {
-        match self {
-            Executable::Native(nf) => Some(nf(avm, ac, this, args)),
-            Executable::Action(af) => {
-                avm.insert_stack_frame_for_function(af.swf_version, af.data.clone(), this, ac);
-
-                for i in 0..args.len() {
-                    avm.push(args.get(i).unwrap().clone());
-                    if let Some(argname) = af.params.get(i) {
-                        avm.current_stack_frame_mut().unwrap().define(argname, args.get(i).unwrap().clone(), ac.gc_context);
-                    }
-                }
-
-                None
-            }
-        }
-    }
-}
 
 pub const TYPE_OF_OBJECT: &str = "object";
 pub const TYPE_OF_FUNCTION: &str = "function";
